--- downloaders/config.py
@@ -1,5 +1,0 @@
-TELEGRAM_API_ID = '107649'
-TELEGRAM_API_HASH = '4c0cbb403eb99d14e4ba38a4e40c2e13'
-TELEGRAM_PHONE = '+380986558186'
-
-TELEGRAM_USER_DIALOG_MESSAGES_LIMIT = None
